requirements.txt
psycopg2-binary==2.9.10
flask==3.1.1
